updated some internals

+++ MCClone/WorldStorage.cs
@@ -6,14 +6,14 @@ namespace MCClone
 {
     public class Block
     {
-        public Block(int X, int Y, int Z)
+        public Block(int X, UInt16 Y, int Z)
         {
             this.X = X;
             this.Y = Y;
             this.Z = Z;
         }
         public int X { get; set; }
-        public int Y { get; set; }
+        public UInt16 Y { get; set; }
         public int Z { get; set; }
     }
     public class Chunk
